requirements.txt
firebase_functions~=0.1.0
firebase-admin==6.5.0
requests>=2.31.0
google-cloud-firestore==2.11.1
